+++ apps/backend/flow-manager/CHANGELOG.md
@@ -3,3 +3,7 @@
 ## [0.0.1] - 2026-01-06
 
 - First commit
+
+## [0.0.2] - 2026-01-31
+
+- flow - verify-email completed